feat(location remove): added location remove functionality

+++ app/api/users/me/locations/[id]/route.ts
@@ -7,6 +7,8 @@ import {Location} from "@prisma/client";
 import {LocationSchema} from "@/utils/validators";
 import {transformValidationErrors} from "@/utils/functions";
 import NotAllowed from "@/errors/not-allowed";
+import PictureRepository from "@/repositories/picture.repository";
+import {removeLocationImageFolder} from "@/utils/fs";
 
 interface Params {
     params: { id: Location["id"] }
@@ -40,14 +42,42 @@ export async function GET(request: NextRequest, {params}: Params) {
 }
 
 export async function DELETE(request: Request, {params}: Params) {
-    // todo: check if user is the owner, remove imgs and then delete the location
+    const userId = await getUserIdFromRequest();
     const {id} = params;
 
-    await LocationRepository.delete(id);
+    try {
+        const location = await LocationRepository.getWithReservations(id);
 
-    return NextResponse.json({
-        message: "Location deleted"
-    } satisfies ResponseDTO<never>);
+        if (userId !== location.userId) {
+            return NextResponse.json({
+                message: "User is not the location's owner"
+            } satisfies ResponseDTO<never>, {
+                status: 403
+            });
+        }
+
+        if (location.reservations.length > 0) {
+            return NextResponse.json({
+                message: "Cannot delete location, has reservations"
+            } satisfies ResponseDTO<never>, {
+                status: 406
+            });
+        }
+
+        await removeLocationImageFolder(id);
+        await PictureRepository.deleteManyByLocation(id);
+        await LocationRepository.delete(id);
+
+        return NextResponse.json({
+            message: "Location deleted"
+        } satisfies ResponseDTO<never>);
+    } catch (e: any) {
+        return NextResponse.json({
+            message: e.message || "Server error"
+        } satisfies ResponseDTO<never>, {
+            status: e.statusCode || 500
+        });
+    }
 }
 
 export async function PATCH(request: Request, {params}: Params) {
